requirements.txt
Django==1.5.2
distribute==0.6.34
dj-database-url==0.2.2
dj-static==0.0.5
django-geoposition==0.1.4
gunicorn==17.5
static==0.4
wsgiref==0.1.2
